+++ R/fill_missing_interval.R
@@ -9,7 +9,8 @@
 #' @export
 #'
 #' @examples
-#' fill_missing_strict(c(NA, 1))
+#' fill_missing_interval(c(NA, 1, 2, NA))
+#' fill_missing_interval(c(NA, 10, 20, NA))
 fill_missing_interval <- function(x, min_known_n = NULL, min_known_p = NULL) {
     fill_missing(x, min_known_n, min_known_p, type = "interval")
 }